There no actual code changes in this commit except for the change in #include All changes are related to moving files around and organizing them in directories
