+++ CMakeLists.txt
@@ -53,38 +53,38 @@ endif()
 add_executable(${PROJECT_NAME} src/main.cpp
         src/Game.h
         src/Game.cpp
-        src/RectangleMeshComponent.h
-        src/RectangleMeshComponent.cpp
-        src/Actor.h
-        src/Actor.cpp
-        src/Cell.h
-        src/Cell.cpp
-        src/Component.h
-        src/Component.cpp
-        src/Grid.h
-        src/Grid.cpp
-        src/DrawComponent.cpp
-        src/DrawComponent.h
-        src/TextComponent.cpp
-        src/TextComponent.h
-        src/LineMeshComponent.cpp
-        src/LineMeshComponent.h
-        src/Animator.cpp
-        src/Animator.h
-        src/Animation.cpp
-        src/Animation.h
-        src/TransformAnimation.cpp
-        src/TransformAnimation.h
-        src/FlashingAnimation.cpp
-        src/FlashingAnimation.h
-        src/SpriteComponent.cpp
-        src/SpriteComponent.h
-        src/BoxCollider.cpp
-        src/BoxCollider.h
-        src/BackButton.cpp
-        src/BackButton.h
-        src/Logo.cpp
-        src/Logo.h
+        src/components/RectangleMeshComponent.h
+        src/components/RectangleMeshComponent.cpp
+        src/actors/Actor.h
+        src/actors/Actor.cpp
+        src/actors/Cell.h
+        src/actors/Cell.cpp
+        src/components/Component.h
+        src/components/Component.cpp
+        src/actors/Grid.h
+        src/actors/Grid.cpp
+        src/components/DrawComponent.cpp
+        src/components/DrawComponent.h
+        src/components/TextComponent.cpp
+        src/components/TextComponent.h
+        src/components/LineMeshComponent.cpp
+        src/components/LineMeshComponent.h
+        src/components/animator/Animator.cpp
+        src/components/animator/Animator.h
+        src/components/animator/Animation.cpp
+        src/components/animator/Animation.h
+        src/components/animator/TransformAnimation.cpp
+        src/components/animator/TransformAnimation.h
+        src/components/animator/FlashingAnimation.cpp
+        src/components/animator/FlashingAnimation.h
+        src/components/SpriteComponent.cpp
+        src/components/SpriteComponent.h
+        src/components/BoxCollider.cpp
+        src/components/BoxCollider.h
+        src/actors/BackButton.cpp
+        src/actors/BackButton.h
+        src/actors/Logo.cpp
+        src/actors/Logo.h
 )
 #set(raylib_VERBOSE 1)
 target_link_libraries(${PROJECT_NAME} raylib)
