+++ src/components/LineMeshComponent.cpp
