+++ src/components/TextComponent.cpp
